move manifest building types and files into manifest_building folder.
