+++ src/wrapped_radix_engine_toolkit/high_level/mod.rs
@@ -1,15 +1,9 @@
-#[macro_use]
-mod assert_manifest;
-
-mod manifests;
-mod manifests_create_tokens;
+mod manifest_building;
 mod ret_api;
 mod sargon_specific_types;
 mod token_definition_metadata;
 
-pub use assert_manifest::*;
-pub use manifests::*;
-pub use manifests_create_tokens::*;
+pub use manifest_building::*;
 pub use ret_api::*;
 pub use sargon_specific_types::*;
 pub use token_definition_metadata::*;
